maiecho_py/pyproject.toml
[project]
name = "maiecho-py"
version = "0.1.0"
description = "Python 3.12 scaffold for the MaiEcho service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15,<2.0",
    "apscheduler>=3.10,<4.0",
    "fastapi>=0.116,<1.0",
    "httpx>=0.28,<1.0",
    "jinja2>=3.1,<4.0",
    "openai>=1.68,<2.0",
    "psutil>=7.0,<8.0",
    "pydantic-settings>=2.8,<3.0",
    "pyyaml>=6.0,<7.0",
    "sqlalchemy>=2.0,<3.0",
    "tenacity>=9.0,<10.0",
    "uvicorn>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.15,<2.0",
    "pytest>=8.3,<9.0",
    "ruff>=0.11,<1.0",
    "types-psutil>=7.0,<8.0",
    "types-pyyaml>=6.0,<7.0",
]

[project.scripts]
maiecho-py = "maiecho_py.cmd.maiecho.main:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
maiecho_py = ["py.typed", "config/*.yaml", "prompts/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["maiecho_py"]

[[tool.mypy.overrides]]
module = ["apscheduler", "apscheduler.*"]
ignore_missing_imports = true

[tool.basedpyright]
include = ["src", "tests"]
extraPaths = ["src"]
venvPath = "."
venv = ".venv"
